+++ package.json
@@ -110,6 +110,7 @@
     "@types/webpack-env": "^1.18.4",
     "@typescript-eslint/eslint-plugin": "^7.1.1",
     "@typescript-eslint/parser": "^7.1.1",
+    "@vitest/coverage-v8": "^1.3.1",
     "approvals": "^6.2.4",
     "aws-sdk-client-mock": "^3.0.1",
     "c8": "^9.1.0",
