+++ test/base-compiler-tests.ts
@@ -662,9 +662,9 @@ Args: []
             '/tmp/somefolder/tmp/someotherfolder/test.h',
         );
 
-        // TODO(partouf); I (mgodbolt) had to disable this as I wasn't sure what it was trying to do:
-        // https://github.com/compiler-explorer/compiler-explorer/blob/d392f130934febe3c76cd2a47dbe062716cf1107/test/base-compiler-tests.ts
-        // expect(compiler.getExtraFilepath('/tmp/somefolder', '\\test.h')).toEqual('/tmp/somefolder/test.h');
+        if (process.platform === 'win32') {
+            expect(compiler.getExtraFilepath('/tmp/somefolder', '\\test.h')).toEqual('/tmp/somefolder/test.h');
+        }
 
         expect(() => compiler.getExtraFilepath('/tmp/somefolder', 'test_hello/../../etc/passwd')).toThrow(Error);
