+++ vitest.config.ts
@@ -5,6 +5,7 @@ import {defineConfig} from 'vitest/config';
 export default defineConfig({
     test: {
         include: ['test/**/*.ts'],
+        exclude: ['test/_*.ts', 'test/utils.ts'],
         setupFiles: ['/test/_setup-fake-aws.ts', '/test/_setup-log.ts'],
     },
 });
